+++ src/core/generator/config/defaultConfig.ts
@@ -5,7 +5,7 @@ import type { RecursivePartial } from '../../types/recursivePartial';
 /**
  * The options control the generator and decide which icons are disabled or not.
  */
-export const getDefaultConfiguration = (): Required<Config> => ({
+export const getDefaultConfig = (): Required<Config> => ({
   folders: {
     theme: 'specific',
     color: '#90a4ae',
@@ -33,10 +33,7 @@ export const getDefaultConfiguration = (): Required<Config> => ({
 export const padWithDefaultConfig = (
   config?: RecursivePartial<Config>
 ): Config => {
-  const withDefaultConfig = merge(
-    getDefaultConfiguration(),
-    config ?? {}
-  ) as Config;
+  const withDefaultConfig = merge(getDefaultConfig(), config ?? {}) as Config;
 
   return withDefaultConfig;
 };
