+++ src/core/helpers/configHash.ts
@@ -1,4 +1,4 @@
-import { getDefaultConfiguration } from '../generator/config/defaultConfig';
+import { getDefaultConfig } from '../generator/config/defaultConfig';
 import { logger } from '../logging/logger';
 import type { Config } from '../models/icons/config';
 
@@ -8,7 +8,7 @@ import type { Config } from '../models/icons/config';
  */
 export const getFileConfigHash = (config: Partial<Config>): string => {
   try {
-    const defaults = getDefaultConfiguration();
+    const defaults = getDefaultConfig();
     let fileConfigString = '';
     if (
       config.saturation !== defaults.saturation ||
