+++ src/extension/commands/saturation.ts
@@ -1,6 +1,6 @@
 import { window as codeWindow } from 'vscode';
 import {
-  getDefaultConfiguration,
+  getDefaultConfig,
   logger,
   translate,
   validateSaturationValue,
@@ -40,7 +40,7 @@ const validateSaturationInput = (saturationInput: string) => {
 
 /** Get the current value of the saturation of the icons. */
 export const getCurrentSaturationValue = (): number => {
-  const defaultConfig = getDefaultConfiguration();
+  const defaultConfig = getDefaultConfig();
   return getThemeConfig<number>('saturation') ?? defaultConfig.saturation;
 };
