+++ src/scripts/icons/generateJson.ts
@@ -6,12 +6,12 @@ import {
   generateFileIcons,
   generateFolderIcons,
   generateManifest,
-  getDefaultConfiguration,
+  getDefaultConfig,
 } from '../../core';
 
 try {
   // Generate default file and folder icons
-  const config = getDefaultConfiguration();
+  const config = getDefaultConfig();
   generateFileIcons(config.files.color, config.opacity, config.saturation);
   generateFolderIcons(config.folders.color, config.opacity, config.saturation);
